+++ test/utils/kratos/send-kratos-flow.ts
@@ -1,0 +1,10 @@
+import request from 'supertest';
+
+export const sendKratosFlow = async (
+  actionUrl: string,
+  params: Record<string, unknown>
+) =>
+  request(actionUrl)
+    .post('')
+    .set('Accept', 'application/json')
+    .send(params);
